% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat-helpers.R
\name{clean_countries}
\alias{clean_countries}
\title{clean_countries}
\usage{
clean_countries(
  df,
  code_col = "area_code",
  year_col = "year",
  value_col = "value"
)
}
\arguments{
\item{df}{a dataframe containing the correct columns}

\item{code_col}{column name containing faostat codes}

\item{year_col}{column name containing years as \%Y format}

\item{value_col}{column name containing the faostat values}
}
\value{
dataframe with double counted and missing values removed
}
\description{
Cleans the countries imported from FAOSTAT by removing multiple versions
of China for any one year and removing countries which invalid ISO3 codes
}
