% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-functions.R
\name{generate_kbl}
\alias{generate_kbl}
\title{generate_kbl}
\usage{
generate_kbl(
  df,
  col_names,
  caption,
  header_spec_fun,
  footnotes,
  output_name = NULL
)
}
\arguments{
\item{df}{the tables data}

\item{col_names}{column names}

\item{caption}{table header}

\item{header_spec_fun}{a function which will be applied to the column header}

\item{footnotes}{a character vector of footnotes to be added to the table}

\item{output_name}{the output location to save the kable}
}
\description{
Generates a kable using the kable extra package and saves the output
pdf
}
