% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{get_aquaculture_value}
\alias{get_aquaculture_value}
\title{get_aquaculture_value}
\usage{
get_aquaculture_value(
  aquaculture_value,
  value_col = constant_2014_2016_usd_value,
  years = c(1998, 2003, 2008, 2013, 2018)
)
}
\arguments{
\item{years}{}
}
\description{
get_aquaculture_value
}
