% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{get_crop_value}
\alias{get_crop_value}
\title{get_crop_value}
\usage{
get_crop_value(
  crop_value,
  value_col = gross_production_value_constant_2014_2016_usd,
  years = c(1998, 2003, 2008, 2013, 2018)
)
}
\arguments{
\item{crop_value}{a dataframe containing the crop values}

\item{years}{}
}
\description{
Returns the crop values for the specified years
}
