% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat-helpers.R
\name{get_gbads_file}
\alias{get_gbads_file}
\title{get_gbads_file}
\usage{
get_gbads_file(str, dir)
}
\arguments{
\item{str}{regex string to match file name}

\item{dir}{directory of gbads file}
}
\value{
the full path the the appropriate name
}
\description{
get_gbads_file
}
