% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat-helpers.R
\name{get_livestock_codes}
\alias{get_livestock_codes}
\title{get_livestock_codes}
\usage{
get_livestock_codes(
  code_dir = config$data$codes$faostat$dir,
  write_codes = TRUE
)
}
\arguments{
\item{code_dir}{directory containing rds files of FAOSTAT item codes}
}
\value{
a vector containing the livestock codes used
}
\description{
Returns the FAO item codes which will be used in the
valuation of livestock types
}
