% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{get_livestock_output_value}
\alias{get_livestock_output_value}
\title{get_livestock_output_value}
\usage{
get_livestock_output_value(
  livestock_value,
  value_col = gross_production_value_constant_2014_2016_usd,
  years = c(1998, 2003, 2008, 2013, 2018),
  in_cat = c("cattle", "chicken", "pig", "sheep"),
  out_cat = "Other Livestock"
)
}
\description{
get_livestock_output_value
}
