% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{nature_color_scheme}
\alias{nature_color_scheme}
\title{nature_color_scheme
Obtained from \code{ggsci::pal_npg()(7)}
Visualise through \code{scales::show_col(ggsci::pal_npg()(7))}}
\usage{
nature_color_scheme()
}
\value{
nature colorscheme
}
\description{
nature_color_scheme
Obtained from \code{ggsci::pal_npg()(7)}
Visualise through \code{scales::show_col(ggsci::pal_npg()(7))}
}
