% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faostat-helpers.R
\name{sanitize_columns}
\alias{sanitize_columns}
\title{sanitize_columns}
\usage{
sanitize_columns(df, exclude = c("iso3_code"))
}
\arguments{
\item{df}{a dataframe}

\item{exclude}{character vector containing which column names will be excluded}
}
\value{
a dataframe which only has ASCII characters in its character columns
}
\description{
this function sanitizes character columns by removing unusual characters
and turns selected colums to snakecase
}
\examples{
sanitize_columns(data.frame(a = "a"))
}
