% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{scale_trill}
\alias{scale_trill}
\title{scale_trill}
\usage{
scale_trill(x, accuracy = 0.5)
}
\arguments{
\item{x}{the numeric vector to be scaled}

\item{accuracy}{what accuracy to display the results at}
}
\description{
Scales axis to trillions of dollars
}
