% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write_pretty_json}
\alias{write_pretty_json}
\title{write_pretty_json}
\usage{
write_pretty_json(x, file_name)
}
\arguments{
\item{x}{the list object to be serialised}

\item{file_name}{file name of output json}
}
\description{
Serialises a list to a json file with
pretty indentations to make it easier
to read
}
\examples{

t <- tempfile()
x <- list(a = "a", b = "b")
write_pretty_json(x, t)
cat(paste(readLines(t), collapse = "\n"))

}
